提交可以编译多个平台；支持简单命令控制

+++ live/liveMedia/include/Mp4DirectorySource.hh
@@ -73,7 +73,7 @@ private:
   unsigned fLastPlayTime;
   Boolean fHaveStartedReading;
   Boolean fLimitNumBytesToStream;
-  unsigned int fWriteLen;
+  size_t fWriteLen;
   u_int64_t fNumBytesToStream; // used iff "fLimitNumBytesToStream" is True
   videoMap fVideos;
   videoMap::iterator fVideoIterator;
